chore(file source): refactor file source into common module
