+++ lib/file-source-common/src/metadata_ext.rs
