+++ lib/file-source/src/file_watcher/tests/experiment_no_truncations.rs
@@ -1,12 +1,10 @@
 use std::{fs, io::Write};
 
 use bytes::Bytes;
+use file_source_common::ReadFrom;
 use quickcheck::{QuickCheck, TestResult};
 
-use crate::{
-    file_watcher::{tests::*, FileWatcher, RawLineResult},
-    ReadFrom,
-};
+use crate::file_watcher::{tests::*, FileWatcher, RawLineResult};
 
 // Interpret all FWActions, excluding truncation
 //
