+++ lib/file-source/src/paths_provider.rs
@@ -1,12 +1,36 @@
-//! [`Glob`] paths provider.
+//! [`Glob`] based paths provider implementation.
 
 use std::path::PathBuf;
 
 pub use glob::MatchOptions;
 use glob::Pattern;
 
-use super::PathsProvider;
-use crate::FileSourceInternalEvents;
+use file_source_common::internal_events::FileSourceInternalEvents;
+
+/// Represents the ability to enumerate paths.
+///
+/// For use in [`crate::file_server::FileServer`].
+///
+/// # Notes
+///
+/// Ideally we'd use an iterator with bound lifetime here:
+///
+/// ```ignore
+/// type Iter<'a>: Iterator<Item = PathBuf> + 'a;
+/// fn paths(&self) -> Self::Iter<'_>;
+/// ```
+///
+/// However, that's currently unavailable at Rust.
+/// See: <https://github.com/rust-lang/rust/issues/44265>
+///
+/// We use an `IntoIter` here as a workaround.
+pub trait PathsProvider {
+    /// Provides the iterator that returns paths.
+    type IntoIter: IntoIterator<Item = PathBuf>;
+
+    /// Provides a set of paths.
+    fn paths(&self) -> Self::IntoIter;
+}
 
 /// A glob-based path provider.
 ///
