+++ lib/vector-lib/Cargo.toml
@@ -9,6 +9,7 @@ publish = false
 codecs = { path = "../codecs", default-features = false }
 enrichment = { path = "../enrichment" }
 file-source = { path = "../file-source", optional = true }
+file-source-common = { path = "../file-source-common", optional = true }
 opentelemetry-proto = { path = "../opentelemetry-proto", optional = true }
 prometheus-parser = { path = "../prometheus-parser", optional = true }
 vector-api-client = { path = "../vector-api-client", optional = true }
@@ -25,7 +26,7 @@ vrl = { workspace = true, optional = true }
 api = ["vector-tap/api"]
 api-client = ["dep:vector-api-client"]
 lua = ["vector-core/lua"]
-file-source = ["dep:file-source"]
+file-source = ["dep:file-source", "dep:file-source-common"]
 opentelemetry = ["dep:opentelemetry-proto"]
 prometheus = ["dep:prometheus-parser"]
 proptest = ["vector-lookup/proptest", "vrl/proptest"]
