+++ lib/vector-lib/src/lib.rs
@@ -2,6 +2,8 @@ pub use codecs;
 pub use enrichment;
 #[cfg(feature = "file-source")]
 pub use file_source;
+#[cfg(feature = "file-source")]
+pub use file_source_common;
 #[cfg(feature = "api-client")]
 pub use vector_api_client as api_client;
 pub use vector_buffers as buffers;
